Very cool SSTV signal generator
For now, uses glReadPixels to get the current framebuffer and exports a Martin 2 signal in araw series of floats to a file to be read in by Audacity or whatever else can read it.

There's some timing issues somewhere, but I'm really shocked at how quickly this came together. Go me! :)