SSTV code refactor, many more modes
The SSTV code was refactored to be a lot more modular, and as such there are a number of new modes added. Instructions now have flags that denote state about line indexing and pitch/length control, as well as modes gaining new fields for better control.

VIS is now calculated by hand, which means new modes could be added. As such, the new modes are: Martin 1, Wraase SC2-180, and Pasokon P3, P5, and P7.