Lights + LightManager, position meshes correctly (temp)
More hack-fixes, this time useful.

A Light struct was added, which allows lights to be more dynamic than just their position in a vec3 array. Doing this allows a LightManager to update the uniforms for a specific shader program by piping in its data.

Meshes that are part of models also have their own unique transforms within the model. In lieu of getting the scene graph properly set up, meshes now have a transform matrix that is an offset from the main model matrix.